divided route and header. added old svg files
